[fix] #77 출시 전 상품 삭제 로직 수정
product.isDone()을 product.isReleased()로 수정하여 출시 상태를 명확히 확인하도록 수정

+++ src/main/java/com/nuclear/boomm/product/service/ProductService.java
@@ -125,6 +125,11 @@ public class ProductService {
     public ProductResponse deleteUnReleasedProduct(Long productId) {
         Product product = productRepository.findByProductId(productId)
                 .orElseThrow(() -> new CustomException(ErrorCode.PRODUCT_NOT_FOUND));
+
+        if (product.isReleased()) {
+            throw new CustomException(ErrorCode.PRODUCT_IS_RELEASED);
+        }
+
         productRepository.delete(product);
 
         return ProductResponse.from(product);
